app/pyproject.toml
[tool.poetry]
name = "szalasapp"
version = "0.1.0"
description = "Equipment and malfunction management system with Firebase backend"
authors = ["lodyczekoladowe <65414380+lodyczekoladowe@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.2"
firebase-admin = "^7.1.0"
python-dotenv = "^1.1.1"
pandas = "^2.3.3"
google-cloud-recaptcha-enterprise = "^1.29.0"
google-cloud-firestore = "^2.21.0"
google-cloud-storage = "^3.4.1"
qrcode = "^8.2"
python-docx = "^1.1.2"
openpyxl = "^3.1.5"
reportlab = "^4.2.5"
pillow = "^12.0.0"
httpx = {extras = ["http2"], version = "^0.28.1"}
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
authlib = "^1.6.6"
requests = "^2.32.5"

[tool.poetry.group.prod.dependencies]
gunicorn = "^23.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-flask = "^1.3.0"
black = "^25.1.0"
flake8 = "^7.1.1"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
